+++ packages/reactflux-react/src/index.ts
@@ -1,2 +1,4 @@
+export * from './types'
 export { useStore } from './useStore'
+export { useDevtools } from './useDevtools'
 export type { Selector } from './types'